chore: refactor styles and settingsHome
