+++ domains/Settings/SettingsHome/AccountSettings/Password/index.js
@@ -1,7 +1,7 @@
 import { Parse } from 'parse/react-native';
 import React, { useState } from 'react';
 import {
-  ActivityIndicator, Alert, StyleSheet, View
+  ActivityIndicator, Alert, View
 } from 'react-native';
 import {
   Button, Headline, Text, TextInput
@@ -10,6 +10,7 @@ import {
 import { getData, storeData } from '../../../../../modules/async-storage';
 import I18n from '../../../../../modules/i18n';
 import { theme } from '../../../../../modules/theme';
+import styles from '../../../index.styles'
 
 const Password = () => {
   const [submitting, setSubmitting] = useState(false);
@@ -99,54 +100,10 @@ const Password = () => {
           color={theme.colors.primary}
         />
       ) : (
-        <Button mode="contained" onPress={() => changePassword()}>{I18n.t('passwordSettings.newPassword')}</Button>
+        <Button mode="contained" onPress={() => changePassword()}>{I18n.t('passwordSettings.changePassword')}</Button>
       )}
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  mainContainer: {
-    paddingLeft: '5%',
-    paddingRight: '5%'
-  },
-  textContainer: {
-    flexDirection: 'row'
-  },
-  buttonContainer: {
-    flexDirection: 'row',
-    marginLeft: 'auto',
-    marginTop: 'auto',
-    marginBottom: 'auto',
-    flex: 1
-  },
-  lineContainer: {
-    marginBottom: 30,
-  },
-  svg: {
-    marginLeft: 'auto',
-    marginTop: -3,
-    marginBottom: -5
-  },
-  text: {
-    flexShrink: 1,
-    // fontWeight: 'bold',
-    fontSize: 16,
-    color: '#555',
-    marginVertical: 7,
-  },
-  horizontalLineGray: {
-    borderBottomColor: '#D0D0D0',
-    borderBottomWidth: 1,
-    marginTop: 10,
-    marginBottom: 10,
-  },
-  horizontalLinePrimary: {
-    borderBottomColor: theme.colors.primary,
-    borderBottomWidth: 1,
-    marginTop: 10,
-    marginBottom: 10,
-  }
-});
 
 export default Password;
