+++ domains/Settings/SettingsHome/AccountSettings/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {
-  StyleSheet, View
+  View
 } from 'react-native';
 import {
   Button
@@ -11,73 +11,35 @@ import Language from './Language';
 import NamePhoneEmail from './NamePhoneEmail';
 import Password from './Password';
 
+import styles from '../../index.styles';
+
 const AccountSettings = ({
   accountSettingsView, setAccountSettingsView
 }) => (
-  <View style={styles.mainContainer}>
-    <View>
+    <View style={styles.mainContainer}>
       {accountSettingsView === 'NamePhoneEmail' && (
         <NamePhoneEmail />
       )}
-    </View>
-    {
-      accountSettingsView === 'ChangePassword' && (
+      {accountSettingsView === 'ChangePassword' && (
         <Password />
       )
-    }
-    {
-      accountSettingsView === 'FindRecords' && (
+      }
+      {accountSettingsView === 'FindRecords' && (
         <FindRecords />
       )
-    }
-    {
-      accountSettingsView === 'Language' && (
+      }
+      {accountSettingsView === 'Language' && (
         <Language />
       )
-    }
-    <Button onPress={() => {
-      setAccountSettingsView('');
-    }}
-    >
-      Back
+      }
+      <Button onPress={() => {
+        setAccountSettingsView('');
+      }}
+      >
+        Back
     </Button>
 
-  </View>
-);
-
-const styles = StyleSheet.create({
-  mainContainer: {
-    paddingLeft: '5%',
-    paddingRight: '5%'
-  },
-  textContainer: {
-    flexDirection: 'row'
-  },
-  buttonContainer: {
-    flexDirection: 'row',
-    marginLeft: 'auto',
-    marginTop: 'auto',
-    marginBottom: 'auto',
-    flex: 1
-  },
-  svg: {
-    marginLeft: 'auto',
-    marginTop: -3,
-    marginBottom: -5
-  },
-  text: {
-    flexShrink: 1,
-    // fontWeight: 'bold',
-    fontSize: 16,
-    color: '#555',
-    marginVertical: 7,
-  },
-  horizontalLineGray: {
-    borderBottomColor: '#D0D0D0',
-    borderBottomWidth: 1,
-    marginTop: 10,
-    marginBottom: 10,
-  }
-});
+    </View>
+  );
 
 export default AccountSettings;
