+++ domains/Settings/SettingsHome/index.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import {
-  StyleSheet, View
+  View
 } from 'react-native';
 import {
   Button, Headline, IconButton, Text
@@ -10,10 +10,30 @@ import I18n from '../../../modules/i18n';
 import { theme } from '../../../modules/theme';
 import AccountSettings from './AccountSettings';
 
+import styles from '../index.styles';
+
 const SettingsHome = ({
   setView, prevView, logOut, settingsView, setSettingsView
 }) => {
   const [accountSettingsView, setAccountSettingsView] = useState('');
+  const inputs = [
+    {
+      key: 'NamePhoneEmail',
+      label: I18n.t('accountSettings.namePhoneEmail')
+    },
+    {
+      key: 'ChangePassword',
+      label: I18n.t('accountSettings.changePassword')
+    },
+    {
+      key: 'FindRecords',
+      label: I18n.t('accountSettings.findRecords')
+    },
+    {
+      key: 'Language',
+      label: I18n.t('accountSettings.language')
+    }
+  ];
 
   return (
     <View>
@@ -25,64 +45,29 @@ const SettingsHome = ({
             </View>
             {/* <View style={{ paddingLeft: '5%' }}>
               <Button onPress={() => setSettingsView('Support')}>
-              {I18n.t('accountSettings.support')}</Button>
+                {I18n.t('accountSettings.support')}</Button>
             </View> */}
           </View>
           <View style={{ paddingLeft: '5%', paddingRight: '5%', paddingTop: 20 }}>
             <Headline style={{ fontWeight: 'bold' }}>{I18n.t('accountSettings.accountSettings')}</Headline>
             <View style={styles.horizontalLineGray} />
-            <View style={{ flexDirection: 'row' }}>
-              <Text style={styles.text}>{I18n.t('accountSettings.namePhoneEmail')}</Text>
-              <IconButton
-                icon="chevron-right"
-                size={30}
-                color={theme.colors.primary}
-                style={{ marginLeft: 'auto', marginTop: -5, marginBottom: -10 }}
-                onPress={() => {
-                  setAccountSettingsView('NamePhoneEmail');
-                }}
-              />
-            </View>
-            <View style={styles.horizontalLineGray} />
-            <View style={{ flexDirection: 'row' }}>
-              <Text style={styles.text}>{I18n.t('accountSettings.changePassword')}</Text>
-              <IconButton
-                icon="chevron-right"
-                size={30}
-                color={theme.colors.primary}
-                style={{ marginLeft: 'auto', marginTop: -5, marginBottom: -10 }}
-                onPress={() => {
-                  setAccountSettingsView('ChangePassword');
-                }}
-              />
-            </View>
-            <View style={styles.horizontalLineGray} />
-            <View style={{ flexDirection: 'row' }}>
-              <Text style={styles.text}>{I18n.t('accountSettings.findRecords')}</Text>
-              <IconButton
-                icon="chevron-right"
-                size={30}
-                color={theme.colors.primary}
-                style={{ marginLeft: 'auto', marginTop: -5, marginBottom: -10 }}
-                onPress={() => {
-                  setAccountSettingsView('FindRecords');
-                }}
-              />
-            </View>
-            <View style={styles.horizontalLineGray} />
-            <View style={{ flexDirection: 'row' }}>
-              <Text style={styles.text}>{I18n.t('accountSettings.language')}</Text>
-              <IconButton
-                icon="chevron-right"
-                size={30}
-                color={theme.colors.primary}
-                style={{ marginLeft: 'auto', marginTop: -5, marginBottom: -10 }}
-                onPress={() => {
-                  setAccountSettingsView('Language');
-                }}
-              />
-            </View>
-            <View style={styles.horizontalLineGray} />
+            {inputs.length && inputs.map((input) => (
+              <View>
+                <View style={{ flexDirection: 'row' }}>
+                  <Text style={styles.text}>{input.label}</Text>
+                  <IconButton
+                    icon="chevron-right"
+                    size={30}
+                    color={theme.colors.primary}
+                    style={{ marginLeft: 'auto', marginTop: -5, marginBottom: -10 }}
+                    onPress={() => {
+                      setAccountSettingsView(input.key);
+                    }}
+                  />
+                </View>
+                <View style={styles.horizontalLineGray} />
+              </View>
+            ))}
           </View>
           <Button onPress={() => {
             setView(prevView);
@@ -90,7 +75,7 @@ const SettingsHome = ({
           >
             {I18n.t('accountSettings.back')}
           </Button>
-          <Button mode="contained" onPress={logOut} style={{ marginTop: 20 }}>{I18n.t('accountSettings.logout')}</Button>
+          <Button mode="contained" onPress={logOut} style={{ marginTop: 20, marginLeft: '5%', marginRight: '5%' }}>{I18n.t('accountSettings.logout')}</Button>
         </View>
       )}
       {accountSettingsView !== '' && (
@@ -106,39 +91,5 @@ const SettingsHome = ({
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  cardSmallStyle: {
-    height: 110,
-    width: 150,
-    marginHorizontal: 7,
-    marginVertical: 7,
-  },
-  svg: {
-    marginLeft: 'auto',
-    marginRight: 'auto',
-  },
-  cardContainer: {
-    alignItems: 'center',
-    marginHorizontal: 14,
-    marginVertical: 14,
-  },
-  textContainer: {
-    flexDirection: 'row'
-  },
-  text: {
-    flexShrink: 1,
-    // fontWeight: 'bold',
-    fontSize: 16,
-    color: '#555',
-    marginVertical: 7,
-  },
-  horizontalLineGray: {
-    borderBottomColor: '#D0D0D0',
-    borderBottomWidth: 1,
-    marginTop: 10,
-    marginBottom: 10,
-  },
-});
 
 export default SettingsHome;
